apps/agents/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
anthropic>=0.20.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
